rename multiprocessing module to avoid clash with standard lib module which led to a crash in pylint
